Clarify cast to int

+++ defaultlook.cpp
@@ -1787,7 +1787,7 @@ void defaultlook::setupBrightness()
 {
     //get brightness value for currently shown display
     QString brightness=runCmd("LANG=C xrandr --verbose | awk '/" + ui->comboBoxDisplay->currentText() +"/{flag=1;next}/Clones/{flag=0}flag'|grep Brightness|cut -d' ' -f2").output;
-    int brightness_slider_value = brightness.toFloat() * 100;
+    int brightness_slider_value = static_cast<int>(brightness.toFloat() * 100);
     ui->horizontalSliderBrightness->setValue(brightness_slider_value);
     if (verbose) qDebug() << "brightness string is " << brightness;
     if (verbose) qDebug() << " brightness_slider_value is " << brightness_slider_value;
